requirements.txt
torch==2.7.1
torchvision
omegaconf
transformers==4.56.1
timm
open_clip_torch
einops
scipy
pillow
accelerate
gdown
huggingface-hub
torch_fidelity
torchinfo
webdataset
xformers
flash-attn>=2.0.0  # Optional: For FlashAttention2/3 support (requires CUDA and compilation)
bitsandbytes
tensorflow-cpu
tensorflow-datasets